ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
chromadb
gitpython
python-dotenv
pydantic
requests
python-multipart
google-generativeai
tiktoken
